@@ -13,12 +13,12 @@ class MealItem extends StatelessWidget {
   // final Function removeItem;
 
   MealItem({
-    @required this.id,
-    @required this.title,
-    @required this.imageUrl,
-    @required this.affordability,
-    @required this.complexity,
-    @required this.duration,
+    required this.id,
+    required this.title,
+    required this.imageUrl,
+    required this.affordability,
+    required this.complexity,
+    required this.duration,
     // @required this.removeItem
   });
 
