@@ -22,10 +22,10 @@ class _FiltersScreenState extends State<FiltersScreen> {
 
   @override
   initState() {
-    _glutenFree = widget.currentFilters['gluten'];
-    _lactoseFree = widget.currentFilters['lactose'];
-    _vegetarian = widget.currentFilters['vegetarian'];
-    _vegan = widget.currentFilters['vegan'];
+    _glutenFree = widget.currentFilters['gluten']!;
+    _lactoseFree = widget.currentFilters['lactose']!;
+    _vegetarian = widget.currentFilters['vegetarian']!;
+    _vegan = widget.currentFilters['vegan']!;
     super.initState();
   }
 
@@ -33,7 +33,7 @@ class _FiltersScreenState extends State<FiltersScreen> {
     String title,
     String description,
     bool currentValue,
-    Function updateValue,
+      Function(bool?) updateValue,
   ) {
     return SwitchListTile(
       title: Text(title),
@@ -85,7 +85,7 @@ class _FiltersScreenState extends State<FiltersScreen> {
                   (newValue) {
                     setState(
                       () {
-                        _glutenFree = newValue;
+                        _glutenFree = newValue!;
                       },
                     );
                   },
@@ -97,7 +97,7 @@ class _FiltersScreenState extends State<FiltersScreen> {
                   (newValue) {
                     setState(
                       () {
-                        _lactoseFree = newValue;
+                        _lactoseFree = newValue!;
                       },
                     );
                   },
@@ -109,7 +109,7 @@ class _FiltersScreenState extends State<FiltersScreen> {
                   (newValue) {
                     setState(
                       () {
-                        _vegetarian = newValue;
+                        _vegetarian = newValue!;
                       },
                     );
                   },
@@ -121,7 +121,7 @@ class _FiltersScreenState extends State<FiltersScreen> {
                   (newValue) {
                     setState(
                       () {
-                        _vegan = newValue;
+                        _vegan = newValue!;
                       },
                     );
                   },
